FIXED 2 bugs
1.  For some reason the confirm-removal-prompt css/html started
appearing messed up, as somehow the width of the column changed and
allowed the green button to rise to the 2nd line of the span.  Since
the button has a big top-margin, it would mess up the layout
completely.  Solution was to add a <br> at the end of the span.

2. Instead of using “N, N’ as empty values for attire and val for
product option variation, which causes a bug as no alphanumeric are
allowed to be passed as parameters through an HTML onclick=.  Using ‘0,
0’ instead.